Move JS files into ./js directory
