+++ .gitignore
@@ -9,3 +9,4 @@ android/.idea
 android/build
 android/app/build
 ios/
+.vscode/
