+++ index.ios.js
@@ -1,3 +1,3 @@
 import { AppRegistry } from 'react-native';
-import App from './App';
+import App from './js/App';
 AppRegistry.registerComponent('mtglifecounter', () => App);
