+++ js/components/NumberSpinner.test.js
