+++ js/screens/Game/GameSettings.js
