+++ js/screens/Game/GameSettings.test.js
